requirements.txt
tensorflow==2.13.0
numpy
scikit-learn
keras
pickle5
streamlit
pillow